+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/JdbcBatchWriteDao.java
@@ -1,19 +1,15 @@
 package com.dwarfeng.subgrade.impl.dao;
 
-import com.dwarfeng.subgrade.sdk.jdbc.mapper.CrudMapper;
-import com.dwarfeng.subgrade.sdk.jdbc.template.CreateTableTemplate;
-import com.dwarfeng.subgrade.sdk.jdbc.template.CrudTemplate;
+import com.dwarfeng.subgrade.sdk.jdbc.JdbcBatchWriteProcessor;
+import com.dwarfeng.subgrade.sdk.jdbc.SQLAndParameter;
 import com.dwarfeng.subgrade.stack.bean.entity.Entity;
 import com.dwarfeng.subgrade.stack.dao.BatchWriteDao;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.lang.NonNull;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 /**
  * 使用 Jdbc 实现的 BatchWriteDao。
@@ -25,81 +21,61 @@ import java.util.stream.Collectors;
 @SuppressWarnings("DuplicatedCode")
 public class JdbcBatchWriteDao<E extends Entity<?>> implements BatchWriteDao<E> {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(JdbcBatchBaseDao.class);
+    private JdbcTemplate template;
+    private JdbcBatchWriteProcessor<E> processor;
 
-    private JdbcTemplate jdbcTemplate;
-    private CrudTemplate crudTemplate;
-    private CrudMapper<?, E> crudMapper;
-
-    public JdbcBatchWriteDao(
-            @NonNull JdbcTemplate jdbcTemplate,
-            @NonNull CrudTemplate crudTemplate,
-            @NonNull CrudMapper<?, E> crudMapper) {
-        this(jdbcTemplate, crudTemplate, crudMapper, null);
-    }
-
-    public JdbcBatchWriteDao(
-            @NonNull JdbcTemplate jdbcTemplate,
-            @NonNull CrudTemplate crudTemplate,
-            @NonNull CrudMapper<?, E> crudMapper,
-            CreateTableTemplate createTableTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.crudTemplate = crudTemplate;
-        this.crudMapper = crudMapper;
-        if (Objects.nonNull(createTableTemplate)) {
-            creatTable(createTableTemplate);
-        }
-    }
-
-    private void creatTable(CreateTableTemplate createTableTemplate) {
-        List<String> crateTableSQLList = createTableTemplate.createTableSQL();
-        for (String crateTableSQL : crateTableSQLList) {
-            jdbcTemplate.execute(crateTableSQL);
-        }
+    public JdbcBatchWriteDao(@NonNull JdbcTemplate template, @NonNull JdbcBatchWriteProcessor<E> processor) {
+        this.template = template;
+        this.processor = processor;
     }
 
     @Override
-    public void write(E entity) throws DaoException {
+    public void write(E element) throws DaoException {
         try {
-            String insertSQL = crudTemplate.insertSQL();
-            jdbcTemplate.update(insertSQL, crudMapper.insert2Args(entity));
+            internalWrite(element);
         } catch (Exception e) {
             throw new DaoException(e);
         }
     }
 
+    private void internalWrite(E element) {
+        SQLAndParameter sqlAndParameter = processor.provideWrite(element);
+        template.update(sqlAndParameter.getSql(), sqlAndParameter.getParameters());
+    }
+
     @Override
-    public void batchWrite(List<E> entities) throws DaoException {
+    public void batchWrite(List<E> elements) throws DaoException {
         try {
-            String insertSQL = crudTemplate.insertSQL();
-            jdbcTemplate.batchUpdate(insertSQL,
-                    entities.stream().map(crudMapper::insert2Args).collect(Collectors.toList()));
+            if (processor.loopWrite()) {
+                for (E element : elements) {
+                    internalWrite(element);
+                }
+            } else {
+                SQLAndParameter sqlAndParameter = processor.provideBatchWrite(elements);
+                if (Objects.nonNull(sqlAndParameter.getParametersList())) {
+                    template.batchUpdate(sqlAndParameter.getSql(), sqlAndParameter.getParametersList());
+                } else {
+                    template.update(sqlAndParameter.getSql(), sqlAndParameter.getParameters());
+                }
+            }
         } catch (Exception e) {
             throw new DaoException(e);
         }
     }
 
-    public JdbcTemplate getJdbcTemplate() {
-        return jdbcTemplate;
+    public JdbcTemplate getTemplate() {
+        return template;
     }
 
-    public void setJdbcTemplate(@NonNull JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
+    public void setTemplate(JdbcTemplate template) {
+        this.template = template;
     }
 
-    public CrudTemplate getCrudTemplate() {
-        return crudTemplate;
+    public JdbcBatchWriteProcessor<E> getProcessor() {
+        return processor;
     }
 
-    public void setCrudTemplate(@NonNull CrudTemplate crudTemplate) {
-        this.crudTemplate = crudTemplate;
-    }
-
-    public CrudMapper<?, E> getCrudMapper() {
-        return crudMapper;
-    }
-
-    public void setCrudMapper(@NonNull CrudMapper<?, E> crudMapper) {
-        this.crudMapper = crudMapper;
+    public void setProcessor(JdbcBatchWriteProcessor<E> processor) {
+        this.processor = processor;
     }
 }