+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/JdbcWriteDao.java
@@ -1,18 +1,12 @@
 package com.dwarfeng.subgrade.impl.dao;
 
-import com.dwarfeng.subgrade.sdk.jdbc.mapper.CrudMapper;
-import com.dwarfeng.subgrade.sdk.jdbc.template.CreateTableTemplate;
-import com.dwarfeng.subgrade.sdk.jdbc.template.CrudTemplate;
+import com.dwarfeng.subgrade.sdk.jdbc.JdbcWriteProcessor;
+import com.dwarfeng.subgrade.sdk.jdbc.SQLAndParameter;
 import com.dwarfeng.subgrade.stack.bean.entity.Entity;
 import com.dwarfeng.subgrade.stack.dao.WriteDao;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.lang.NonNull;
-
-import java.util.List;
-import java.util.Objects;
 
 /**
  * 使用 Jdbc 实现的 WriteDao。
@@ -24,70 +18,37 @@ import java.util.Objects;
 @SuppressWarnings("DuplicatedCode")
 public class JdbcWriteDao<E extends Entity<?>> implements WriteDao<E> {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(JdbcBatchBaseDao.class);
+    private JdbcTemplate template;
+    private JdbcWriteProcessor<E> processor;
 
-    private JdbcTemplate jdbcTemplate;
-    private CrudTemplate crudTemplate;
-    private CrudMapper<?, E> crudMapper;
-
-    public JdbcWriteDao(
-            @NonNull JdbcTemplate jdbcTemplate,
-            @NonNull CrudTemplate crudTemplate,
-            @NonNull CrudMapper<?, E> crudMapper) {
-        this(jdbcTemplate, crudTemplate, crudMapper, null);
-    }
-
-    public JdbcWriteDao(
-            @NonNull JdbcTemplate jdbcTemplate,
-            @NonNull CrudTemplate crudTemplate,
-            @NonNull CrudMapper<?, E> crudMapper,
-            CreateTableTemplate createTableTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.crudTemplate = crudTemplate;
-        this.crudMapper = crudMapper;
-        if (Objects.nonNull(createTableTemplate)) {
-            creatTable(createTableTemplate);
-        }
-    }
-
-    private void creatTable(CreateTableTemplate createTableTemplate) {
-        List<String> crateTableSQLList = createTableTemplate.createTableSQL();
-        for (String crateTableSQL : crateTableSQLList) {
-            jdbcTemplate.execute(crateTableSQL);
-        }
+    public JdbcWriteDao(@NonNull JdbcTemplate template, @NonNull JdbcWriteProcessor<E> processor) {
+        this.template = template;
+        this.processor = processor;
     }
 
     @Override
-    public void write(E entity) throws DaoException {
+    public void write(E element) throws DaoException {
         try {
-            String insertSQL = crudTemplate.insertSQL();
-            jdbcTemplate.update(insertSQL, crudMapper.insert2Args(entity));
+            SQLAndParameter sqlAndParameter = processor.provideWrite(element);
+            template.update(sqlAndParameter.getSql(), sqlAndParameter.getParameters());
         } catch (Exception e) {
             throw new DaoException(e);
         }
     }
 
-    public JdbcTemplate getJdbcTemplate() {
-        return jdbcTemplate;
+    public JdbcTemplate getTemplate() {
+        return template;
     }
 
-    public void setJdbcTemplate(@NonNull JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
+    public void setTemplate(JdbcTemplate template) {
+        this.template = template;
     }
 
-    public CrudTemplate getCrudTemplate() {
-        return crudTemplate;
+    public JdbcWriteProcessor<E> getProcessor() {
+        return processor;
     }
 
-    public void setCrudTemplate(@NonNull CrudTemplate crudTemplate) {
-        this.crudTemplate = crudTemplate;
-    }
-
-    public CrudMapper<?, E> getCrudMapper() {
-        return crudMapper;
-    }
-
-    public void setCrudMapper(@NonNull CrudMapper<?, E> crudMapper) {
-        this.crudMapper = crudMapper;
+    public void setProcessor(JdbcWriteProcessor<E> processor) {
+        this.processor = processor;
     }
 }