+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 #### 功能构建
 
 - com.dwarfeng.subgrade.sdk.jdbc 包结构调整。
-  - 该调整将会导致此版本与 Release_1.1.0_20200611_build_A 不兼容。
+  - **该调整将会导致此版本与 Release_1.1.0_20200611_build_A 不兼容。**
 - 更改 README.md 的`推荐使用版本`条目，并且移除 git 中 Release_1.1.0_20200611_build_A 的版本标签。
 - 添加 MyBatis 数据访问层支持。
   - com.dwarfeng.subgrade.impl.dao.MyBatisBaseDao
@@ -16,7 +16,17 @@
   - com.dwarfeng.subgrade.impl.dao.MyBatisBatchWriteDao
   - com.dwarfeng.subgrade.impl.dao.MyBatisEntireLookupDao
   - com.dwarfeng.subgrade.impl.dao.MyBatisPresetLookupDao
-
+- 对 JDBC 数据访问层支持进行大规模修改。
+  - **该调整将会导致此版本与 Release_1.1.0_20200611_build_A 不兼容。**
+  - com.dwarfeng.subgrade.impl.dao.JdbcBaseDao
+  - com.dwarfeng.subgrade.impl.dao.JdbcBatchBaseDao
+  - com.dwarfeng.subgrade.impl.dao.JdbcRelationDao
+  - com.dwarfeng.subgrade.impl.dao.JdbcBatchRelationDao
+  - com.dwarfeng.subgrade.impl.dao.JdbcBatchWriteDao
+  - com.dwarfeng.subgrade.impl.dao.JdbcBatchWriteDao
+  - com.dwarfeng.subgrade.impl.dao.JdbcEntireLookupDao
+  - com.dwarfeng.subgrade.impl.dao.JdbcPresetLookupDao
+  
 #### Bug修复
 
 - 修复使用jdbc框架进行 Apache Phoenix 分页查询时生成错误的 SQL 语句的bug。